# Add a new previous_health column in service table