--- a/api/external/applications/applications.swagger.json
+++ b/api/external/applications/applications.swagger.json
@@ -297,6 +297,9 @@
         },
         "site": {
           "type": "string"
+        },
+        "previous_health_check": {
+          "$ref": "#/definitions/applicationsHealthStatus"
         }
       }
     },