--- a/api/external/applications/applications.swagger.json
+++ b/api/external/applications/applications.swagger.json
@@ -260,9 +260,11 @@
         "OK",
         "WARNING",
         "CRITICAL",
-        "UNKNOWN"
+        "UNKNOWN",
+        "NONE"
       ],
       "default": "OK",
+      "description": "- NONE: The representation of NO health check status\nTODO @afiune how much effort would be to change\nthe OK enum to be NONE",
       "title": "The HealthStatus enum matches the habitat implementation for health-check status:\n=\u003e https://www.habitat.sh/docs/reference/#health-check"
     },
     "applicationsService": {