--- a/components/applications-service/pkg/server/server.go
+++ b/components/applications-service/pkg/server/server.go
@@ -233,16 +233,17 @@ func convertStorageServicesToApplicationsServices(svcs []*storage.Service) []*ap
 	services := make([]*applications.Service, len(svcs))
 	for i, svc := range svcs {
 		services[i] = &applications.Service{
-			SupervisorId: svc.SupMemberID,
-			Release:      svc.FullReleaseString(),
-			Group:        svc.Group,
-			HealthCheck:  applications.HealthStatus(applications.HealthStatus_value[svc.Health]),
-			Status:       applications.ServiceStatus(applications.ServiceStatus_value[svc.Status]),
-			Application:  svc.Application,
-			Environment:  svc.Environment,
-			Fqdn:         svc.Fqdn,
-			Channel:      svc.Channel,
-			Site:         svc.Site,
+			SupervisorId:        svc.SupMemberID,
+			Release:             svc.FullReleaseString(),
+			Group:               svc.Group,
+			HealthCheck:         applications.HealthStatus(applications.HealthStatus_value[svc.Health]),
+			Status:              applications.ServiceStatus(applications.ServiceStatus_value[svc.Status]),
+			Application:         svc.Application,
+			Environment:         svc.Environment,
+			Fqdn:                svc.Fqdn,
+			Channel:             svc.Channel,
+			Site:                svc.Site,
+			PreviousHealthCheck: applications.HealthStatus(applications.HealthStatus_value[svc.PreviousHealth]),
 		}
 	}
 	return services