--- a/components/applications-service/pkg/storage/client.go
+++ b/components/applications-service/pkg/storage/client.go
@@ -48,6 +48,7 @@ type Service struct {
 	Environment         string
 	Channel             string
 	Site                string
+	PreviousHealth      string    `db:"previous_health"`
 	LastEventOccurredAt time.Time `db:"last_event_occurred_at"`
 }