--- a/components/applications-service/pkg/storage/postgres/schema/sql/05_previous_health_service.up.sql
+++ b/components/applications-service/pkg/storage/postgres/schema/sql/05_previous_health_service.up.sql
@@ -1,0 +1,2 @@
+-- Add previous_health column
+ALTER TABLE service ADD COLUMN previous_health TEXT NOT NULL DEFAULT '';