--- a/components/applications-service/pkg/storage/postgres/schema/sql/05_previous_health_service.up.sql
+++ b/components/applications-service/pkg/storage/postgres/schema/sql/05_previous_health_service.up.sql
@@ -1,2 +1,6 @@
 -- Add previous_health column
 ALTER TABLE service ADD COLUMN previous_health TEXT NOT NULL DEFAULT '';
+
+-- (fix) drop the default value for the package_ident column
+-- @afiune: we do not want to allow this field to be empty
+ALTER TABLE service ALTER COLUMN package_ident DROP DEFAULT;