--- a/components/applications-service/pkg/storage/postgres/schema/sql/05_previous_health_service.up.sql
+++ b/components/applications-service/pkg/storage/postgres/schema/sql/05_previous_health_service.up.sql
@@ -1,5 +1,6 @@
--- Add previous_health column
-ALTER TABLE service ADD COLUMN previous_health TEXT NOT NULL DEFAULT '';
+-- Add previous_health column with the default HealthStatus 'NONE' that matches
+-- with the protobuf definition => api/external/applications/applications.proto
+ALTER TABLE service ADD COLUMN previous_health TEXT NOT NULL DEFAULT 'NONE';
 
 -- (fix) drop the default value for the package_ident column
 -- @afiune: we do not want to allow this field to be empty