--- a/components/applications-service/pkg/storage/postgres/service.go
+++ b/components/applications-service/pkg/storage/postgres/service.go
@@ -24,6 +24,7 @@ SELECT id
   , channel
   , package_ident
   , last_event_occurred_at
+  , previous_health
 FROM service
 WHERE name = $1
   AND sup_id IN (
@@ -47,6 +48,7 @@ SELECT s.id
   , s.channel as channel
   , sup.site as site
   , s.last_event_occurred_at as last_event_occurred_at
+  , s.previous_health as previous_health
 FROM service AS s
 LEFT JOIN service_group AS sg
   ON s.group_id = sg.id
@@ -73,6 +75,7 @@ SELECT s.id
   , s.channel as channel
   , sup.site as site
   , s.last_event_occurred_at as last_event_occurred_at
+  , s.previous_health as previous_health
 FROM service AS s
 LEFT JOIN service_group AS sg
   ON s.group_id = sg.id