--- a/components/automate-gateway/api/applications.pb.swagger.go
+++ b/components/automate-gateway/api/applications.pb.swagger.go
@@ -300,6 +300,9 @@ func init() {
         },
         "site": {
           "type": "string"
+        },
+        "previous_health_check": {
+          "$ref": "#/definitions/applicationsHealthStatus"
         }
       }
     },