… api/external/applications/applications.pb.go
… benchmarks/applications-service/txt/PR_693.txt