### :nut_and_bolt: Description
Save the previous health check status in the database to power the time wizard
from [JIRA#A2-665](https://chefio.atlassian.net/browse/A2-665)

### :+1: Definition of Done
We are now storing the previous health check status inside the database.

### :athletic_shoe: Demo Script / Repro Steps
Build and start both, the `applications-service` and `automate-gateway`.
Populate the database with some data and check that the previous status is saved.

A service now should look like:
```
    {
      "supervisorId": "aee0210f-7847-4c74-a010-e1805ac3f527",
      "release": "custom/nginx/1.0.1/20190115184823",
      "group": "nginx.kansas-prod",
      "healthCheck": "CRITICAL",
      "application": "demo",
      "environment": "demo",
      "fqdn": "localhost",
      "channel": "stable",
      "site": "testsite",
      "previousHealthCheck": "NONE"
    }
```

Update service(s) health check status and verify that the status is saved! 

### :chains: Related Resources
https://chefio.atlassian.net/browse/A2-895

### :white_check_mark: Checklist

- [ ] Necessary tests added/updated?
- [ ] Necessary docs added/updated?
- [ ] Code actually executed?
- [ ] Vetting performed (unit tests, lint, etc.)?