@afiune 2019-06-28
Just added the benchmarks to compare it between previous PRs, the only change to
consider is that we increased the memory consumption and allocations count which
is expected since we are storing timestamps and calculating previous states:
```
[7][default:/src:0]# benchstat /src/benchmarks/applications-service/txt/PR_580.txt /src/benchmarks/applications-service/txt/PR_693.txt
name                                                   old time/op    new time/op    delta
IngestionNewServices-4                                   7.91ms ± 4%    8.17ms ± 7%     ~     (p=0.690 n=5+5)
IngestionUpdateServices-4                                2.83ms ± 8%    2.91ms ± 6%     ~     (p=0.151 n=5+5)
GetServiceGroups/max_1000_with_0_groups-4                2.04ms ± 5%    1.99ms ± 2%     ~     (p=0.690 n=5+5)
GetServiceGroups/max_1000_with_100_groups-4              11.1ms ± 3%     8.1ms ± 7%  -27.37%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_200_groups-4              27.8ms ± 1%    16.9ms ± 1%  -39.23%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_300_groups-4              52.4ms ± 1%    30.5ms ± 1%  -41.87%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_400_groups-4              85.2ms ± 1%    47.9ms ± 1%  -43.73%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_500_groups-4               125ms ± 0%      70ms ± 2%  -43.79%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_600_groups-4               174ms ± 0%      96ms ± 1%  -44.70%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_700_groups-4               146ms ±60%     122ms ± 1%     ~     (p=0.690 n=5+5)
GetServiceGroups/max_1000_with_800_groups-4               158ms ± 3%     158ms ± 2%     ~     (p=0.690 n=5+5)
GetServiceGroups/max_1000_with_900_groups-4               195ms ± 1%     196ms ± 0%     ~     (p=0.548 n=5+5)
GetServiceGroups/max_1000_with_1000_groups-4              236ms ± 3%     238ms ± 0%     ~     (p=0.151 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_0_groups-4       2.01ms ± 2%    1.92ms ± 3%   -4.42%  (p=0.008 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_1000_groups-4     228ms ± 2%     229ms ± 1%     ~     (p=0.690 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_2000_groups-4     892ms ± 5%     878ms ± 1%     ~     (p=0.310 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_3000_groups-4     1.96s ± 1%     1.95s ± 3%     ~     (p=0.548 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_4000_groups-4     3.47s ± 1%     3.51s ± 2%     ~     (p=0.222 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_5000_groups-4     5.49s ± 3%     5.57s ±14%     ~     (p=0.548 n=5+5)
GetServices/max_1000_with_0_services-4                   2.09ms ± 3%    2.15ms ± 1%   +3.08%  (p=0.016 n=5+5)
GetServices/max_1000_with_100_services-4                 2.62ms ± 3%    2.64ms ± 2%     ~     (p=0.841 n=5+5)
GetServices/max_1000_with_200_services-4                 2.84ms ± 3%    2.92ms ± 1%   +2.99%  (p=0.032 n=5+5)
GetServices/max_1000_with_300_services-4                 3.17ms ± 3%    3.21ms ± 0%     ~     (p=0.190 n=5+4)
GetServices/max_1000_with_400_services-4                 3.19ms ± 6%    3.26ms ± 1%     ~     (p=0.151 n=5+5)
GetServices/max_1000_with_500_services-4                 3.43ms ± 2%    3.53ms ± 1%   +2.81%  (p=0.008 n=5+5)
GetServices/max_1000_with_600_services-4                 3.65ms ± 1%    3.82ms ± 1%   +4.79%  (p=0.008 n=5+5)
GetServices/max_1000_with_700_services-4                 3.87ms ± 2%    4.14ms ± 3%   +6.99%  (p=0.008 n=5+5)
GetServices/max_1000_with_800_services-4                 4.17ms ± 2%    4.38ms ± 2%   +4.95%  (p=0.008 n=5+5)
GetServices/max_1000_with_900_services-4                 4.42ms ± 4%    4.75ms ± 2%   +7.43%  (p=0.008 n=5+5)
GetServices/max_1000_with_1000_services-4                4.65ms ± 2%    4.87ms ± 1%   +4.81%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_0_services-4          1.77ms ± 4%    1.77ms ± 6%     ~     (p=0.841 n=5+5)
GetServicesHeavyLoad/max_5000_with_1000_services-4       4.71ms ± 3%    4.90ms ± 2%   +3.94%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_2000_services-4       7.31ms ± 1%    7.86ms ± 8%   +7.52%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_3000_services-4       10.0ms ± 2%    10.8ms ± 6%   +7.94%  (p=0.016 n=5+5)
GetServicesHeavyLoad/max_5000_with_4000_services-4       12.5ms ± 2%    13.5ms ± 1%   +8.24%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_5000_services-4       15.0ms ± 1%    15.9ms ± 2%   +5.82%  (p=0.008 n=5+5)

name                                                   old alloc/op   new alloc/op   delta
IngestionNewServices-4                                   16.7kB ± 0%    20.1kB ± 0%  +19.89%  (p=0.008 n=5+5)
IngestionUpdateServices-4                                10.6kB ± 0%    13.8kB ± 0%  +29.80%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_0_groups-4                8.87kB ± 0%    8.87kB ± 0%     ~     (p=1.000 n=5+5)
GetServiceGroups/max_1000_with_100_groups-4              40.6kB ± 0%    40.6kB ± 0%   -0.02%  (p=0.000 n=4+5)
GetServiceGroups/max_1000_with_200_groups-4              40.6kB ± 0%    40.6kB ± 0%     ~     (p=0.127 n=5+4)
GetServiceGroups/max_1000_with_300_groups-4              40.6kB ± 0%    40.6kB ± 0%   -0.02%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_400_groups-4              40.6kB ± 0%    40.6kB ± 0%   -0.06%  (p=0.000 n=5+4)
GetServiceGroups/max_1000_with_500_groups-4              40.7kB ± 0%    40.6kB ± 0%   -0.14%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_600_groups-4              40.7kB ± 0%    40.6kB ± 0%   -0.11%  (p=0.016 n=4+5)
GetServiceGroups/max_1000_with_700_groups-4              41.0kB ± 0%    41.0kB ± 0%   -0.07%  (p=0.029 n=4+4)
GetServiceGroups/max_1000_with_800_groups-4              41.0kB ± 0%    41.1kB ± 0%   +0.10%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_900_groups-4              41.0kB ± 0%    41.0kB ± 0%   +0.08%  (p=0.016 n=4+5)
GetServiceGroups/max_1000_with_1000_groups-4             41.1kB ± 0%    41.2kB ± 0%   +0.18%  (p=0.016 n=4+5)
GetServiceGroupsHeavyLoad/max_5000_with_0_groups-4       8.87kB ± 0%    8.87kB ± 0%     ~     (p=1.000 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_1000_groups-4    41.1kB ± 0%    41.1kB ± 0%     ~     (p=0.175 n=5+4)
GetServiceGroupsHeavyLoad/max_5000_with_2000_groups-4    41.4kB ± 0%    41.4kB ± 0%   +0.04%  (p=0.016 n=5+4)
GetServiceGroupsHeavyLoad/max_5000_with_3000_groups-4    41.9kB ± 0%    41.9kB ± 0%   +0.04%  (p=0.029 n=4+4)
GetServiceGroupsHeavyLoad/max_5000_with_4000_groups-4    42.0kB ± 0%    41.9kB ± 0%     ~     (p=0.405 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_5000_groups-4    41.9kB ± 0%    41.9kB ± 0%   +0.04%  (p=0.029 n=4+4)
GetServices/max_1000_with_0_services-4                   12.3kB ± 0%    14.7kB ± 0%  +19.68%  (p=0.016 n=4+5)
GetServices/max_1000_with_100_services-4                 38.6kB ± 0%    44.7kB ± 0%  +15.60%  (p=0.008 n=5+5)
GetServices/max_1000_with_200_services-4                 38.6kB ± 0%    44.7kB ± 0%  +15.60%  (p=0.008 n=5+5)
GetServices/max_1000_with_300_services-4                 38.6kB ± 0%    44.7kB ± 0%  +15.60%  (p=0.008 n=5+5)
GetServices/max_1000_with_400_services-4                 38.6kB ± 0%    44.7kB ± 0%  +15.61%  (p=0.008 n=5+5)
GetServices/max_1000_with_500_services-4                 38.6kB ± 0%    44.7kB ± 0%  +15.61%  (p=0.008 n=5+5)
GetServices/max_1000_with_600_services-4                 38.6kB ± 0%    44.7kB ± 0%  +15.62%  (p=0.008 n=5+5)
GetServices/max_1000_with_700_services-4                 38.6kB ± 0%    44.7kB ± 0%  +15.62%  (p=0.008 n=5+5)
GetServices/max_1000_with_800_services-4                 38.6kB ± 0%    44.7kB ± 0%  +15.61%  (p=0.008 n=5+5)
GetServices/max_1000_with_900_services-4                 38.6kB ± 0%    44.7kB ± 0%  +15.63%  (p=0.008 n=5+5)
GetServices/max_1000_with_1000_services-4                38.6kB ± 0%    44.7kB ± 0%  +15.60%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_0_services-4          12.3kB ± 0%    14.7kB ± 0%  +19.65%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_1000_services-4       38.6kB ± 0%    44.7kB ± 0%  +15.62%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_2000_services-4       38.6kB ± 0%    44.7kB ± 0%  +15.58%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_3000_services-4       38.6kB ± 0%    44.7kB ± 0%  +15.62%  (p=0.016 n=4+5)
GetServicesHeavyLoad/max_5000_with_4000_services-4       38.6kB ± 0%    44.7kB ± 0%  +15.56%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_5000_services-4       38.7kB ± 0%    44.7kB ± 0%  +15.58%  (p=0.016 n=5+4)

name                                                   old allocs/op  new allocs/op  delta
IngestionNewServices-4                                      602 ± 0%       780 ± 0%  +29.57%  (p=0.008 n=5+5)
IngestionUpdateServices-4                                   458 ± 0%       640 ± 0%  +39.74%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_0_groups-4                   504 ± 0%       504 ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_100_groups-4               1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_200_groups-4               1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_300_groups-4               1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_400_groups-4               1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_500_groups-4               1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_600_groups-4               1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_700_groups-4               1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_800_groups-4               1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_900_groups-4               1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroups/max_1000_with_1000_groups-4              1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroupsHeavyLoad/max_5000_with_0_groups-4          504 ± 0%       504 ± 0%     ~     (all equal)
GetServiceGroupsHeavyLoad/max_5000_with_1000_groups-4     1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroupsHeavyLoad/max_5000_with_2000_groups-4     1.41k ± 0%     1.41k ± 0%     ~     (all equal)
GetServiceGroupsHeavyLoad/max_5000_with_3000_groups-4     1.42k ± 0%     1.42k ± 0%     ~     (all equal)
GetServiceGroupsHeavyLoad/max_5000_with_4000_groups-4     1.42k ± 0%     1.42k ± 0%     ~     (p=0.444 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_5000_groups-4     1.42k ± 0%     1.42k ± 0%     ~     (all equal)
GetServices/max_1000_with_0_services-4                      652 ± 0%       804 ± 0%  +23.31%  (p=0.008 n=5+5)
GetServices/max_1000_with_100_services-4                  1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServices/max_1000_with_200_services-4                  1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServices/max_1000_with_300_services-4                  1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServices/max_1000_with_400_services-4                  1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServices/max_1000_with_500_services-4                  1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServices/max_1000_with_600_services-4                  1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServices/max_1000_with_700_services-4                  1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServices/max_1000_with_800_services-4                  1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServices/max_1000_with_900_services-4                  1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServices/max_1000_with_1000_services-4                 1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_0_services-4             652 ± 0%       804 ± 0%  +23.31%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_1000_services-4        1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_2000_services-4        1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_3000_services-4        1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_4000_services-4        1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
GetServicesHeavyLoad/max_5000_with_5000_services-4        1.48k ± 0%     1.74k ± 0%  +16.98%  (p=0.008 n=5+5)
```